Creating separate file for decorators, protecting category endpoints with API KEY and JWT
