+++ app.py
@@ -26,53 +26,14 @@ def create_app(db_url=None):
     app.config["OPENAPI_SWAGGER_UI_URL"] = "https://cdn.jsdelivr.net/npm/swagger-ui-dist/"
     app.config["SQLALCHEMY_DATABASE_URI"] = db_url or os.getenv("DATABASE_URL", "sqlite:///data.db")
     app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+    app.config["JWT_SECRET_KEY"] = "xxx"
+
     db.init_app(app)
     migrate = Migrate(app, db)
     api = Api(app)
-
-    app.config["JWT_SECRET_KEY"] = "jose"
     jwt = JWTManager(app)
-
-    @jwt.expired_token_loader
-    def expired_token_callback(jwt_header, jwt_payload):
-        return (
-            jsonify({"message": "The token has expired.", "error": "token_expired"}),
-            401,
-        )
-
-    @jwt.invalid_token_loader
-    def invalid_token_callback(error):
-        return (
-            jsonify(
-                {"message": "Signature verification failed.", "error": "invalid_token"}
-            ),
-            401,
-        )
-
-    @jwt.unauthorized_loader
-    def missing_token_callback(error):
-        return (
-            jsonify(
-                {
-                    "description": "Request does not contain an access token.",
-                    "error": "authorization_required",
-                }
-            ),
-            401,
-        )
-
-    @jwt.token_in_blocklist_loader
-    def check_if_token_in_blocklist(jwt_header, jwt_payload):
-        return jwt_payload["jti"] in BLOCKLIST
-
-    @jwt.revoked_token_loader
-    def revoked_token_callback(jwt_header, jwt_payload):
-        return (
-            jsonify(
-                {"description": "The token has been revoked.", "error": "token_revoked"}
-            ),
-            401,
-        )
+    from decorators import expired_token_callback, invalid_token_callback, missing_token_callback, \
+        check_if_token_in_blocklist, revoked_token_callback
 
     api.register_blueprint(ItemBlueprint)
     api.register_blueprint(CategoryBlueprint)
