+++ resources/category.py
@@ -3,22 +3,26 @@ from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError
+from flask_jwt_extended import jwt_required
 
 from db import db
 from models import CategoryModel
 from schemas import CategorySchema
+from decorators import require_api_key
 
 
 blp = Blueprint("Categories", __name__, description="Operations on categories")
 
 
 @blp.route("/category/<int:category_id>")
+@require_api_key
 class Categories(MethodView):
     @blp.response(200, CategorySchema)
     def get(self, category_id):
         category = CategoryModel.query.get_or_404(category_id)
         return category
 
+    @jwt_required
     def delete(self, category_id):
         category = CategoryModel.query.get_or_404(category_id)
         db.session.delete(category)
@@ -27,6 +31,7 @@ class Categories(MethodView):
 
 
 @blp.route("/category")
+@require_api_key
 class CategoriesList(MethodView):
     @blp.response(201, CategorySchema(many=True))
     def get(self):
@@ -34,6 +39,7 @@ class CategoriesList(MethodView):
 
     @blp.arguments(CategorySchema)
     @blp.response(200, CategorySchema)
+    @jwt_required()
     def post(self, category_data):
         category = CategoryModel(**category_data)
         try:
